# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stadium Eats API — in-stadium food ordering for FIFA World Cup Morocco 2030
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stadiums/{stadiumId}/trending:
    get:
      operationId: getTrendingItems
      tags: [feed]
      summary: Trending items in a stadium right now
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
        - name: section
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrendingItem" }
  /stadiums/{stadiumId}/feed:
    get:
      operationId: getStadiumFeed
      tags: [feed]
      summary: Live "fans near you are ordering" feed for a stadium
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FeedEvent" }
  /stadiums/{stadiumId}/match:
    get:
      operationId: getLiveMatch
      tags: [match]
      summary: Live match ticker for a stadium
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LiveMatch" }
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          description: Validation or capacity error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /orders/session/{sessionId}:
    get:
      operationId: listSessionOrders
      tags: [orders]
      summary: List orders for a fan session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /orders/{orderId}/rate:
    post:
      operationId: rateOrder
      tags: [orders]
      summary: Rate a delivered order
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RateOrderBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /rider/login:
    post:
      operationId: riderLogin
      tags: [rider]
      summary: Rider PIN login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RiderLoginBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Rider" }
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /rider/{riderId}/orders:
    get:
      operationId: getRiderOrders
      tags: [rider]
      summary: Orders assigned to a rider
      parameters:
        - name: riderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /rider/{riderId}/stats:
    get:
      operationId: getRiderStats
      tags: [rider]
      summary: A rider's day-of stats (deliveries, avg time)
      parameters:
        - name: riderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiderStats" }
  /orders/{orderId}/status:
    post:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update an order's status
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
    SeatInfo:
      type: object
      required: [stadiumId, section, row, seat]
      properties:
        stadiumId: { type: string }
        section: { type: string }
        row: { type: string }
        seat: { type: string }
...
    OrderItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId: { type: string }
        quantity: { type: integer, minimum: 1 }
    OrderItem:
      type: object
      required: [menuItemId, name, quantity, priceMad]
      properties:
        menuItemId: { type: string }
        name: { type: string }
        quantity: { type: integer }
        priceMad: { type: number }
        imageUrl: { type: string }
    OrderStatus:
      type: string
      enum: [received, preparing, on_the_way, delivered, cancelled]
    Order:
      type: object
      required: [id, sessionId, standId, standName, items, status, seat, subtotalMad, deliveryFeeMad, totalMad, etaMinutes, createdAt, isPreOrder]
      properties:
        id: { type: string }
        sessionId: { type: string }
        standId: { type: string }
        standName: { type: string }
        riderId: { type: string }
        riderName: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        status: { $ref: "#/components/schemas/OrderStatus" }
        seat: { $ref: "#/components/schemas/SeatInfo" }
        subtotalMad: { type: number }
        deliveryFeeMad: { type: number }
        totalMad: { type: number }
        etaMinutes: { type: integer }
        createdAt: { type: string, format: date-time }
        scheduledFor: { type: string, format: date-time }
        deliveredAt: { type: string, format: date-time }
        rating: { type: integer }
        isPreOrder: { type: boolean }
        fanName: { type: string }
    CreateOrderBody:
      type: object
      required: [sessionId, standId, items]
      properties:
        sessionId: { type: string }
        standId: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemInput" }
        scheduledFor: { type: string, format: date-time }
        paymentMethod:
          type: string
          enum: [card, apple_pay, google_pay]
    UpdateOrderStatusBody:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/OrderStatus" }
        riderId: { type: string }
    RateOrderBody:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
    Rider:
      type: object
      required: [id, name, stadiumId, status, activeOrders]
      properties:
        id: { type: string }
        name: { type: string }
        stadiumId: { type: string }
        status:
          type: string
          enum: [available, busy, offline]
        activeOrders: { type: integer }
        avatarUrl: { type: string }
    RiderLoginBody:
      type: object
      required: [pin]
      properties:
        pin: { type: string }
    RiderStats:
      type: object
      required: [riderId, deliveredToday, activeOrders, avgDeliveryMinutes]
      properties:
        riderId: { type: string }
        deliveredToday: { type: integer }
        activeOrders: { type: integer }
        avgDeliveryMinutes: { type: number }
...
    TrendingItem:
      type: object
      required: [menuItemId, name, standId, standName, ordersLastHour, imageUrl, priceMad]
      properties:
        menuItemId: { type: string }
        name: { type: string }
        standId: { type: string }
        standName: { type: string }
        ordersLastHour: { type: integer }
        imageUrl: { type: string }
        priceMad: { type: number }
        section: { type: string }
    FeedEvent:
      type: object
      required: [id, message, section, standName, itemName, createdAt]
      properties:
        id: { type: string }
        message: { type: string }
        section: { type: string }
        standName: { type: string }
        itemName: { type: string }
        imageUrl: { type: string }
        createdAt: { type: string, format: date-time }
    LiveMatch:
      type: object
      required: [stadiumId, homeTeam, awayTeam, homeScore, awayScore, minute, status]
      properties:
        stadiumId: { type: string }
        homeTeam: { type: string }
        awayTeam: { type: string }
        homeFlag: { type: string }
        awayFlag: { type: string }
        homeScore: { type: integer }
        awayScore: { type: integer }
        minute: { type: integer }
        status:
          type: string
          enum: [scheduled, first_half, halftime, second_half, full_time]
        kickoffAt: { type: string, format: date-time }