# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stadium Eats API — in-stadium food ordering for FIFA World Cup Morocco 2030
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderBody {
  sessionId: string;
  standId: string;
  items: OrderItemInput[];
  scheduledFor?: Date;
  paymentMethod?: CreateOrderBodyPaymentMethod;
}