# lib/api-zod/src/generated/types/riderStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stadium Eats API — in-stadium food ordering for FIFA World Cup Morocco 2030
 * OpenAPI spec version: 0.1.0
 */

export interface RiderStats {
  riderId: string;
  deliveredToday: number;
  activeOrders: number;
  avgDeliveryMinutes: number;
}